ai_fastapi/requirements.txt
fastapi==0.135.1
langchain_chroma==1.1.0
langchain_core==1.2.19
langchain_openai==1.1.11
langchain_text_splitters==1.1.1
openai==2.28.0
python-dotenv==1.2.2
Requests==2.32.5
streamlit==1.55.0
uvicorn==0.41.0
